implemet dag schedule
