--- infra/remote/dataproc/versions.tf
@@ -1,8 +1,0 @@
-terraform {
-  required_providers {
-    yandex = {
-      source = "yandex-cloud/yandex"
-    }
-  }
-  required_version = ">= 0.13"
-}
